FredAPI.jl

A Julia interface to the Federal Reserve Economic Data API. This package is issued under the MIT License and is unaffiliated with the Federal Reserve Bank of St. Louis.

This product uses the FRED API but is not endorsed or certified by the Federal Reserve Bank of St. Louis.

Use of the FRED API is subject to Terms of Use, and by using this Julia package to interface with the FRED API, you agree to be bound by the FRED API Terms of Use.

All use of the FRED API requires an API key. You are responsible for requesting and using your own key.

Interface

FredAPI.Responses.CategoryResponse

FredAPI.Responses.ReleasesResponse

FredAPI.category.children

FredAPI.category.get

FredAPI.category.related

FredAPI.category.related_tags

FredAPI.category.series

FredAPI.category.tags

FredAPI.get_api_key

FredAPI.release.dates

FredAPI.release.get

FredAPI.release.related_tags

FredAPI.release.series

FredAPI.release.sources

FredAPI.release.tables

FredAPI.release.tags

FredAPI.releases.dates

FredAPI.releases.get_all

FredAPI.series.categories

FredAPI.series.get

FredAPI.series.observations

FredAPI.series.release

FredAPI.series.search

FredAPI.series.search_related_tags

FredAPI.series.search_tags

FredAPI.series.tags

FredAPI.series.updates

FredAPI.series.vintagedates

FredAPI.set_api_key

FredAPI.source.get

FredAPI.source.get_all

FredAPI.source.releases

FredAPI.tags.get_all

FredAPI.tags.related_tags

FredAPI.tags.series

FredAPI.get_api_key — Method

get_api_key(api_key)

Obtains an api_key for use in HTTP requests

If api_key is an instance of AbstractString, a String representation of api_key is returned.

If api_key is nothing, the global API key is returned. (See set(api_key)(@ref))

This method serves as the default approach to obtaining an api_key in many functions.

FredAPI.set_api_key — Method

set_api_key(api_key)

Sets the global api-key constant to api_key

FredAPI.category.children — Function

children(category_id=0)

Get the children categories for the category_id category

See fred/category/children.

FredAPI.category.get — Function

get(category_id=0)

Get the category for category_id, defaulting to the root category

See fred/category.

FredAPI.category.related — Method

related(category_id)

Get categories related to the category_id category

See fred/category/related.

FredAPI.category.related_tags — Method

related_tags(category_id, tag_names; <keyword arguments>)

Get the related tags for one or more tags within a category

See fred/category/related_tags.

FredAPI.category.series — Method

series(category_id)

Get the series available in the category_id category

See fred/category/series.

FredAPI.category.tags — Method

tags(category_id; <keyword arguments>)

Get the tags for the category_id category

See fred/category/tags.

FredAPI.release.dates — Method

dates(release_id; <keyword arguments>)

Gets release dates for a single data release

See fred/release/dates.

FredAPI.release.get — Method

get(release_id; <keyword arguments>)

Gets a data release

See fred/release.

FredAPI.release.related_tags — Method

related_tags(release_id, tag_names; <keyword arguments>)

Get related tags for tag_names within release release_id

See fred/release/related_tags.

FredAPI.release.series — Method

series(release_id; <keyword arguments>)

Get the series for a data release

See fred/release/series.

FredAPI.release.sources — Method

sources(release_id; <keyword arguments>)

Get the sources for a data release

See fred/release/sources.

FredAPI.release.tables — Method

tables(release_id; <keyword arguments>)

Get release tables for release release_id

See fred/release/tables.

FredAPI.release.tags — Method

tags(release_id; <keyword arguments>)

Get the tags for a release

See fred/release/tags.

FredAPI.releases.dates — Method

dates(; <keyword arguments>)

Get release dates for all releases

See fred/releases/dates.

FredAPI.releases.get_all — Method

get_all(; <keyword arguments>)

Get all releases

See fred/releases.

FredAPI.Responses.CategoryResponse — Type

Returned By:

category

category/children

category/related

series/categories

FredAPI.Responses.ReleasesResponse — Type

Returned By:

releases

FredAPI.series.categories — Method

categories(series_id; <keyword arguments>)

Get the categories for series series_id

See fred/series/categories.

FredAPI.series.get — Method

get(series_id; <keyword arguments>)

Get a data series

See fred/series.

FredAPI.series.observations — Method

observations(series_id; <keyword arguments>)

Get the observations or data values for a data series

See fred/series/observations.

FredAPI.series.release — Method

release(series_id; <keyword arguments>)

Get the release for the series_id series

See fred/series/release.

FredAPI.series.search — Method

search(search_text; <keyword arguments>)

Gets the series that match search_text

See fred/series/search.

FredAPI.series.search_related_tags — Function

search_related_tags(series_search_text, tag_names; <keyword arguments>)

Get the related tags for tag_names matching series_search_text

See fred/series/search/related_tags.

FredAPI.series.search_tags — Method

search_tags(series_search_text; <keyword arguments>)

Get the tags for a series search

See fred/series/search/tags.

FredAPI.series.tags — Method

tags(series_id; <keyword arguments>)

Get tags for the series_id series

See fred/series/tags.

FredAPI.series.updates — Method

updates(; <keyword arguments>)

Get series updates within the last two weeks

See fred/series/updates.

FredAPI.series.vintagedates — Method

vintagedates(series_id; <keyword arguments>)

Get historical dates for when the series_id series data values were released/revised

See fred/series/vintagedates.

FredAPI.source.get — Method

get(source_id; <keyword arguments>)

Get a data source

See fred/source.

FredAPI.source.get_all — Method

get_all(; <keyword arguments>)

Get all data sources

See fred/sources.

FredAPI.source.releases — Method

releases(source_id; <keyword arguments>)

Get the releases for the source_id source

See fred/source/releases.

FredAPI.tags.get_all — Method

get_all(; <keyword arguments>)

Get tags, optionally filtered

See fred/tags.

FredAPI.tags.related_tags — Function

related_tags(tag_names; <keyword arguments>)

Get tags related to tag_names

See fred/related_tags.

FredAPI.tags.series — Function

series(tag_names; <keyword arguments>)

Get the series matching all of tag_names

See fred/tags/series.